온보딩 UI 변경 헤더 통일화 필요 : feat : 온보딩 Step 분리 및 각 Step 정보 Enum 관리 https://github.com/TEAM-ROMROM/RomRom-FE/issues/100
